+++ one_for_all_generator/CHANGELOG.md
@@ -1,4 +1,9 @@
-## Next release
+## 2.0.2
+- build: added support to analyzer to `7.x.x`
+
+## 2.0.1
+- feat: support latest dependencies
+- feat: support to ignore annotation
 - feat: support sealed classes
 - refactor!: renamed PlatformException to PlatformError
 - refactor!: the "Result" objects accept on `error` method a `PlatformError` class